requirements.txt
# FastAPI and API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Streamlit for frontend
streamlit==1.28.1

# LLM and AI dependencies
openai==1.3.6
google-generativeai==0.3.2
langchain==0.0.340
langchain-openai==0.0.2

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3  # Built into Python

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Utilities
typing-extensions==4.8.0
# System monitoring
psutil==5.9.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.7.0

# Optional: For enhanced features
redis==5.0.1  # For caching
celery==5.3.4  # For background tasks